+++ Assessify.Backend/src/main/java/de/assessify/app/assessifyapi/api/dtos/response/GradeDto.java
@@ -3,4 +3,4 @@ package de.assessify.app.assessifyapi.api.dtos.response;
 import java.util.Date;
 import java.util.UUID;
 
-public record GradeDto(UUID id, String gradeName, Float value, Float gradeWeighting, Date date) {}
+public record GradeDto(UUID id, String gradeName, float value, int gradeWeighting, Date date) {}
